Pt 154904411 ws binding address (#740)
* (Naive) configuration of WS binding address

This should allow configuring the websocket binding address using the epoch.yaml file. Instead it crashes the application.

* Fixed the project not starting

thanks to the help of @dincho we made this work

* Enabling users to configure other listen addresses

Not changing the defaults (yet). Updated epoch_full.yaml to showcase the new configuration options.

* Refactored some functions

to get the IP address from configs easier

* Changing default internal listen address

It was 0.0.0.0 before which is not recommended, now it is 127.0.0.1

* Switched back to binary defaults

* Allowing external access to internal port

Since we set the default listening to 127.0.0.1, but in the docker-compose localnet environment we want to have access to the internal port, too.