Pt 157220182 compile dutch auction (#1121)
* Make sure dutch_auction compiles

Note that we don't check the compiled file yet.

* Add sophia compiler tests

These are very simple and currently only test that the contract is
compilable. There is no checking of the output. We have also added a
call to this test module in the CT suite.

* Add bools as ints, true -> 1 and false -> 0

* Fix getting the contract into the compiler

The interface to the compiler is a bit weird so yo need a bit of
trickery.