Introduce efficient TX-pool sync (#1245)
Previously we sent the entire mempool after every ping message - this is
obviuosly not scaling. Here we introduce a proper sync mechanism.
TX-pool synchronization is done between a pair of peers. Since the node
will still work if it miss a transaction or two it is better to keep it
somewhat simple.

The sync is handled by aec_tx_pool_sync which allows for 1 outgoing and
5 incoming simultaneous synchronizations. The synchronization is done by
building an MP tree on each side and only sending the parts of the trees
that differ - thus minimizing the bandwidth used.

* Remove get_mempool
* Bump P2P protocol version + release notes
* Change send_after -> start_timer + fix tests
* Addressing more review comments
* Avoid problems with long running operation (get)
* Export path() type from aeu_mp_trees