Let miner decide whether to include oracle txs, depending on market
Before this change, each oracle tx fee had the following constraints:

* Fee >= "tx generic minimum fee" i.e. `1`
  (`aec_governance:minimum_tx_fee()`)

* Fee >= "oracle txs specific minimum fee" (static), that was in the
  range `[1, 4]` - i.e. greater than (or equal to) 1 - depending on
  the specific oracle tx

* (Fee - "oracle txs specific minimum fee") >= `ceil(TtlDelta *
  0.001)` (dynamic), keeping into account the TTL of objects in the
  state trees

This change brings the minimum static fee component of all oracle txs
(i.e. excluding the TTL dynamic component) to the same as all other
txs.  This means that miners decide which oracle txs to pick depending
on the market.

The minimum tx fee is
[documented](https://github.com/aeternity/protocol/blame/f1e55e5c25aea06ef55721075b820fab3037de04/consensus/consensus.md#L275):

> Fee MUST be at least `1` aeon.