Check value of call instruction against balance in common VM code
... as part of instruction semantics - as opposed to checking it only
in chain callback and primops callbacks.

The aim is introducing a single point where the call value is checked
- both for non-primop and primop calls - for correctness, readability
and maintainability.  The non-primop call is one code path so not
difficult to follow.  Though the primop calls are multiple and not
immediate to follow - potentially not using value, or using it in a
non-straightforward way and returning custom error reason.  Having a
simple point checking this makes the code more straightforward.

A side benefit is that this check is performed also when using
test-only chain callback implementations.  Some tests in
`contract_sophia_bytecode_aevm_SUITE` had to be made more compliant in
order to adapt to these stricter VM checks, and a new negative test in
the same SUITE was introduced for showing this behaviour.

The Sophia compiler is minimally adapted for not ignoring the item
placed on the stack by `?CALL` (symptom was number zero being decoded
as return value in tests).

Introducing internal function `recursive_call2` in 1aevm_eeevm` I
checked that no previously bound variables were used in match patterns
as assertions.