Optimize and cleanup aehttp_contracts_SUITE
Before each contract call was made as a separate transaction + mine two blocks. This takes lots of unnecessary time. New
approach allow for many contract call TXs to go into same micro block. This is done by introducing by keeping track of what TXs
are in mempool, and then mining them in batches. This means we have to save the _check_ on the return value until the TX is
actually on the chain so there is a small mechanism for this introduced. `call_func` is introduced where the last argument is the
delayed check.

Runtime goes from 250s to 90s on my machine.