PT-160256544 calldata creation part (#1709)
* Change mechanism for creating calldata to allow arbitrary Sophia types

The most general way is to set the function name to "" and give a contract
as the argument containing a function __call(). For instance to call a function
foo with arguments arg1 : t1, arg2 : t2, and arg3 : t3, one would write

contract SomeName =
  // ... type defs required for t1, t2, t3
  function foo : (t1, t2, t3) => res
  function __call() = foo(arg1, arg2, arg3)

This will get type checked and appropriate calldata will be generated. A shorthand
is available, if no typedefs are required, where the function name argument is set to
"foo : (t1, t2, t3) => res" and the argument is "arg1, arg2, arg3".

* Disable some debug output

* Handle return type = _ in create_calldata

* Treat hashes longer than 32 bytes as signature literals

* Update aeso_abi_tests

* Update aevm_chain tests

* Update sophia_bytecode tests

* Check that given types match the types in the contract metadata

* Release notes

* Allow address literals for contracts, oracles and oracle queries when building calldata

* Keep old calldata generation and fall back to it if function argument is a single name

* Change http endpoints to take either Fun/Args or CallCode

In the former case it uses the old calldata creation

* Allow hash literals also for old-style calldata creation

* Use new calldata creation in a couple of places in http contracts tests