Add extra nonce handling on server side
Each client is assigned an unique extra nonce upon subscription. The extra
nonce is selected randomly from a range of available nonces determined
by the by size of the nonce.

The extra nonce is saved in a cache in order to keep a track of used nonces
and keep the nonce unique for a client.

This also adds close/1 to server and client session modules to clean
up resources when closing session/connection.